Rename yokogawa subpackage to cellvoyager
