+++ examples/tools/show_FOV_ROIs/show_FOV_ROIs.py
@@ -3,10 +3,10 @@ An example of visualizing FOV ROIs and their overlaps.
 """
 import matplotlib.pyplot as plt
 
-from fractal_tasks_core.lib_ROI_overlaps import run_overlap_check
-from fractal_tasks_core.yokogawa.metadata import (
+from fractal_tasks_core.cellvoyager.metadata import (
     parse_yokogawa_metadata,
 )
+from fractal_tasks_core.lib_ROI_overlaps import run_overlap_check
 
 
 def _plot_rectangle(min_x, min_y, max_x, max_y, overlapping):
