+++ fractal_tasks_core/__init__.py
@@ -1,5 +1,6 @@
 import logging
 
+
 logging.basicConfig(
     level=logging.INFO, format="%(asctime)s; %(levelname)s; %(message)s"
 )
