+++ fractal_tasks_core/cellvoyager/__init__.py
@@ -1,0 +1,3 @@
+"""
+Subpackage with utilities for tasks converting CellVoyager images to OME-Zarr.
+"""
