+++ fractal_tasks_core/cellvoyager/filenames.py
