+++ fractal_tasks_core/cellvoyager/metadata.py
