+++ fractal_tasks_core/tasks/create_ome_zarr.py
@@ -21,6 +21,13 @@ import pandas as pd
 from pydantic.decorator import validate_arguments
 
 import fractal_tasks_core
+from fractal_tasks_core.cellvoyager.filenames import (
+    glob_with_multiple_patterns,
+)
+from fractal_tasks_core.cellvoyager.filenames import parse_filename
+from fractal_tasks_core.cellvoyager.metadata import (
+    parse_yokogawa_metadata,
+)
 from fractal_tasks_core.ome_zarr.channels import check_unique_wavelength_ids
 from fractal_tasks_core.ome_zarr.channels import check_well_channel_labels
 from fractal_tasks_core.ome_zarr.channels import define_omero_channels
@@ -30,11 +37,6 @@ from fractal_tasks_core.roi import prepare_FOV_ROI_table
 from fractal_tasks_core.roi import prepare_well_ROI_table
 from fractal_tasks_core.roi import remove_FOV_overlaps
 from fractal_tasks_core.tables import write_table
-from fractal_tasks_core.yokogawa.filenames import glob_with_multiple_patterns
-from fractal_tasks_core.yokogawa.filenames import parse_filename
-from fractal_tasks_core.yokogawa.metadata import (
-    parse_yokogawa_metadata,
-)
 
 
 __OME_NGFF_VERSION__ = fractal_tasks_core.__OME_NGFF_VERSION__
