--- fractal_tasks_core/yokogawa/__init__.py
@@ -1,3 +1,0 @@
-"""
-Subpackage with utilities for tasks converting Yokogawa images to OME-Zarr.
-"""
