+++ tests/test_unit_parse_metadata_from_filename.py
@@ -14,7 +14,7 @@ Zurich.
 import pytest
 from devtools import debug
 
-from fractal_tasks_core.yokogawa.filenames import parse_filename
+from fractal_tasks_core.cellvoyager.filenames import parse_filename
 
 f1 = (
     "20200812-CardiomyocyteDifferentiation14-Cycle1"
